requirements.txt
typer
httpx
python-dotenv
rich
